moved businessVerificationReducer
